fix: unify assistant visible text sanitizers (#61729)
